+++ CHANGELOG.md
@@ -18,7 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Allow mkforms 10.0.0 (#860)
 - Upgrade to PHP-CS-Fixer V3 (#854)
 - Upgrade to PHPUnit 7 (#853)
-- Require rn_base >= 1.13.12 (#774)
+- Require rn_base >= 1.13.13 (#774, #911)
 - Move PHPStan from PHIVE to Composer (#849)
 - Rename sub-namespace `Interface` to `Interfaces` (#623, #852)